+++ src/api/common/search.ts
@@ -1,63 +1,13 @@
 import * as zod from "zod";
 import { ZodType } from "zod";
 import { clamp, get } from "lodash";
-import { Path, zodPath } from "../../lib/zod/zodPath";
 import { t } from "../trpc";
-
-export const sortDirectionType = zod.union([
-  zod.literal("asc"),
-  zod.literal("desc"),
-]);
-
-export type SortDirection = zod.infer<typeof sortDirectionType>;
-
-export interface SearchQuery<T, F> {
-  filter?: F;
-  sort?: SearchSort<T>;
-  offset?: number;
-  limit?: number;
-}
-
-export type SearchSort<T> = Array<{
-  field: Path<T>;
-  sort: SortDirection;
-}>;
-
-export interface SearchResult<T> {
-  total: number;
-  entities: T[];
-}
-
-export function createSearchTypes<ET extends ZodType, FT extends ZodType>(
-  entityType: ET,
-  filterType: FT
-) {
-  type Entity = zod.infer<ET>;
-  type Filter = zod.infer<FT>;
-
-  const pathType = zodPath(entityType);
-
-  const sortType: ZodType<SearchSort<Entity>> = zod.array(
-    zod.object({
-      field: pathType,
-      sort: sortDirectionType,
-    })
-  );
-
-  const queryType: ZodType<SearchQuery<Entity, Filter>> = zod.object({
-    filter: filterType.optional(),
-    sort: sortType.optional(),
-    offset: zod.number().optional(),
-    limit: zod.number().optional(),
-  });
-
-  const resultType: ZodType<SearchResult<Entity>> = zod.object({
-    total: zod.number(),
-    entities: zod.array(entityType),
-  });
-
-  return { queryType, resultType };
-}
+import {
+  searchTypes,
+  SearchQuery,
+  SearchResult,
+  SearchSort,
+} from "./search.types";
 
 // The default max limit is in place for when the client provides
 // no limit and a search controller has no explicit limit.
@@ -121,7 +71,7 @@ export function createSearchProcedure<ET extends ZodType, FT extends ZodType>(
     filter?: zod.infer<FT>
   ) => number | undefined
 ) {
-  const { queryType, resultType } = createSearchTypes(entityType, filterType);
+  const { queryType, resultType } = searchTypes(entityType, filterType);
   const search = createSearchController(getEntities, isMatch, getMaxLimit);
   return t.procedure
     .input(queryType)
